fix warning

+++ perf-map-file.c
@@ -10,6 +10,6 @@ FILE *perf_map_open(pid_t pid) {
 }
 
 void perf_map_write_entry(FILE *method_file, const void* code_addr, unsigned int code_size, const char* entry) {
-    fprintf(method_file, "%lx %x %s\n", code_addr, code_size, entry);
+    fprintf(method_file, "%lx %x %s\n", (unsigned long) code_addr, code_size, entry);
     fflush(method_file);
 }
